refactor(Calcified): use `Int`s instead of `Byte`s for port numbers in order for better Java compatibility (remove the need for casting to byte which was noisy), resolves #4

+++ src/main/kotlin/dev/frozenmilk/dairy/calcified/CalcifiedDeviceMap.kt
@@ -2,5 +2,5 @@ package dev.frozenmilk.dairy.calcified
 
 import dev.frozenmilk.dairy.calcified.hardware.CalcifiedModule
 
-abstract class CalcifiedDeviceMap<T> internal constructor(protected val module: CalcifiedModule, private val map: MutableMap<Byte, T> = mutableMapOf()) : MutableMap<Byte, T> by map
+abstract class CalcifiedDeviceMap<T> internal constructor(protected val module: CalcifiedModule, private val map: MutableMap<Int, T> = mutableMapOf()) : MutableMap<Int, T> by map
